requirements.txt
fastapi>=0.95.0                # API framework
uvicorn[standard]>=0.22.0       # ASGI server

sqlalchemy>=2.0                # ORM

asyncpg>=0.27.0                 # Postgres driver (remove if using SQLite)
# aiosqlite>=0.17.0              # SQLite async driver (optional)

alembic>=1.11.1               # DB migrations

python-jose[cryptography]>=3.3.0    # JWT token creation/validation
passlib[bcrypt]>=1.7.4          # Password hashing
bcrypt>=4.0.1                  # OS-level bcrypt

pydantic>=1.10.7               # Data validation

celery[redis]>=5.2.7           # Task queue
redis>=4.5.1                  # Redis client

sendgrid>=6.10.0               # Email sending

python-dotenv>=1.0.0            # .env file loader

pytest>=7.3.1                  # Testing (optional)
httpx>=0.24.1                  # HTTP client for testing
